change random functions from simulation.py to utils.py
